refactor: split evolution tests and seams
