+++ forgeflow_runtime/gate_evaluation.py
@@ -1,7 +1,17 @@
 from __future__ import annotations
 
 
+from pathlib import Path
 from typing import Any
+
+from forgeflow_runtime.artifact_validation import (
+    artifact_path,
+    artifact_variants,
+    load_validated_artifact,
+    missing_artifacts,
+)
+from forgeflow_runtime.errors import RuntimeViolation
+from forgeflow_runtime.policy_loader import RuntimePolicy
 
 
 _STAGE_GATE_EVIDENCE_ARTIFACTS = {
@@ -29,3 +39,63 @@ def record_completed_gate(run_state: dict[str, Any], stage_name: str, *, stage_g
     gate_name = stage_gate_map.get(stage_name)
     if gate_name and gate_name not in run_state["completed_gates"]:
         run_state["completed_gates"].append(gate_name)
+
+
+def matching_review_payload(
+    task_dir: Path,
+    expected_review_type: str,
+    expected_verdict: str,
+    *,
+    canonical_task_id: str,
+) -> dict[str, Any] | None:
+    for artifact_name in ["review-report", "review-report-spec", "review-report-quality"]:
+        review_path = artifact_path(task_dir, artifact_name)
+        if not review_path.exists():
+            continue
+        payload = load_validated_artifact(task_dir, artifact_name, expected_task_id=canonical_task_id)
+        _validate_review_semantics(payload, source_name=review_path.name)
+        if payload.get("review_type") == expected_review_type and payload.get("verdict") == expected_verdict:
+            return payload
+    return None
+
+
+def enforce_stage_gate(task_dir: Path, policy: RuntimePolicy, stage_name: str, *, canonical_task_id: str) -> None:
+    gate_name = policy.stage_gate_map.get(stage_name)
+    if gate_name is None:
+        return
+
+    missing_gate_artifacts = missing_artifacts(task_dir, policy.gate_requirements.get(gate_name, []))
+    if missing_gate_artifacts:
+        raise RuntimeViolation(
+            f"{stage_name} requires artifacts satisfying gate {gate_name}: {', '.join(missing_gate_artifacts)}"
+        )
+
+    for required_artifact in policy.gate_requirements.get(gate_name, []):
+        for variant in artifact_variants(required_artifact):
+            variant_path = artifact_path(task_dir, variant)
+            if variant_path.exists():
+                load_validated_artifact(task_dir, variant, expected_task_id=canonical_task_id)
+
+    gate_review = policy.gate_reviews.get(gate_name, {})
+    expected_review_type = gate_review.get("review_type")
+    expected_verdict = gate_review.get("verdict")
+    if expected_review_type and expected_verdict and matching_review_payload(
+        task_dir,
+        expected_review_type,
+        expected_verdict,
+        canonical_task_id=canonical_task_id,
+    ) is None:
+        raise RuntimeViolation(
+            f"{stage_name} requires approved {expected_review_type} review-report artifact"
+        )
+
+
+def _validate_review_semantics(payload: dict[str, Any], *, source_name: str) -> None:
+    verdict = payload.get("verdict")
+    open_blockers = payload.get("open_blockers", [])
+    if open_blockers is None:
+        open_blockers = []
+    if verdict == "approved" and open_blockers:
+        raise RuntimeViolation(f"approved {source_name} cannot declare open_blockers")
+    if verdict == "approved" and payload.get("safe_for_next_stage") is False:
+        raise RuntimeViolation(f"approved {source_name} cannot set safe_for_next_stage=false")
