+++ tests/runtime/test_orchestrator_cli.py
@@ -1,69 +1,12 @@
 import json
-import os
-import subprocess
-import sys
 from pathlib import Path
 
-from forgeflow_runtime.orchestrator import load_runtime_policy, run_route
+from tests.runtime.cli_helpers import ROOT, make_task_dir, run_orchestrator_cli, write_json
 
 
-ROOT = Path(__file__).resolve().parents[2]
-
-
-def _write_json(path: Path, payload: dict) -> None:
-    path.write_text(json.dumps(payload, indent=2), encoding="utf-8")
-
-
-def _make_task_dir(tmp_path: Path) -> Path:
-    task_dir = tmp_path / "task"
-    task_dir.mkdir()
-    _write_json(
-        task_dir / "brief.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "objective": "Run a small route",
-            "in_scope": ["runtime"],
-            "out_of_scope": [],
-            "constraints": ["local only"],
-            "acceptance_criteria": ["route works"],
-            "risk_level": "low",
-        },
-    )
-    _write_json(
-        task_dir / "run-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "current_stage": "clarify",
-            "status": "in_progress",
-            "completed_gates": ["clarification_complete"],
-            "failed_gates": [],
-            "retries": {},
-            "current_task_id": "",
-            "spec_review_approved": False,
-            "quality_review_approved": False,
-        },
-    )
-    return task_dir
-
-
-def _run_orchestrator_cli(
-    *args: str,
-    env: dict[str, str] | None = None,
-    cwd: Path | None = None,
-) -> subprocess.CompletedProcess[str]:
-    effective_env = os.environ.copy()
-    if env:
-        effective_env.update(env)
-    return subprocess.run(
-        [sys.executable, str(ROOT / "scripts" / "run_orchestrator.py"), *args],
-        cwd=cwd or ROOT,
-        capture_output=True,
-        text=True,
-        check=False,
-        env=effective_env,
-    )
+_make_task_dir = make_task_dir
+_run_orchestrator_cli = run_orchestrator_cli
+_write_json = write_json
 
 
 def test_cli_help_includes_operator_shell_examples() -> None:
@@ -328,128 +271,6 @@ def test_cli_supports_recovery_commands(tmp_path: Path) -> None:
     assert json.loads(escalate_result.stdout)["status"] == "blocked"
 
 
-def test_cli_execute_real_codex_uses_binary_from_path_without_live_credentials(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
-    bin_dir = tmp_path / "bin"
-    bin_dir.mkdir()
-    codex = bin_dir / "codex"
-    codex.write_text(
-        "#!/usr/bin/env python3\n"
-        "import sys\n"
-        "assert sys.argv[1] == 'exec'\n"
-        "print('FAKE_CODEX_REAL_OUTPUT')\n",
-        encoding="utf-8",
-    )
-    codex.chmod(0o755)
-
-    result = _run_orchestrator_cli(
-        "execute",
-        "--task-dir",
-        str(task_dir),
-        "--route",
-        "small",
-        "--adapter",
-        "codex",
-        "--real",
-        env={"PATH": f"{bin_dir}:{os.environ.get('PATH', '')}"},
-    )
-
-    assert result.returncode == 0
-    payload = json.loads(result.stdout)
-    assert payload["status"] == "success"
-    assert payload["adapter"] == "codex"
-    assert (task_dir / "clarify-output.md").read_text(encoding="utf-8").strip() == "FAKE_CODEX_REAL_OUTPUT"
-
-
-def test_cli_execute_real_claude_uses_binary_from_path_without_live_credentials(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
-    bin_dir = tmp_path / "bin"
-    bin_dir.mkdir()
-    claude = bin_dir / "claude"
-    claude.write_text(
-        "#!/usr/bin/env python3\n"
-        "import sys\n"
-        "assert '-p' in sys.argv\n"
-        "assert '--bare' in sys.argv\n"
-        "assert '--dangerously-skip-permissions' in sys.argv\n"
-        "print('FAKE_CLAUDE_REAL_OUTPUT')\n",
-        encoding="utf-8",
-    )
-    claude.chmod(0o755)
-
-    result = _run_orchestrator_cli(
-        "execute",
-        "--task-dir",
-        str(task_dir),
-        "--route",
-        "small",
-        "--adapter",
-        "claude",
-        "--real",
-        env={"PATH": f"{bin_dir}:{os.environ.get('PATH', '')}"},
-    )
-
-    assert result.returncode == 0
-    payload = json.loads(result.stdout)
-    assert payload["status"] == "success"
-    assert payload["adapter"] == "claude"
-    assert payload["execution_mode"] == "real"
-    assert (task_dir / "clarify-output.md").read_text(encoding="utf-8").strip() == "FAKE_CLAUDE_REAL_OUTPUT"
-
-
-def test_cli_execute_real_unsupported_adapter_fails_explicitly(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
-
-    result = _run_orchestrator_cli(
-        "execute",
-        "--task-dir",
-        str(task_dir),
-        "--route",
-        "small",
-        "--adapter",
-        "cursor",
-        "--real",
-    )
-
-    assert result.returncode == 0
-    payload = json.loads(result.stdout)
-    assert payload["status"] == "failure"
-    assert payload["error"] == "real adapter unsupported: cursor; supported real adapters: claude, codex"
-
-
-def test_cli_execute_real_codex_missing_binary_is_actionable(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
-
-    result = _run_orchestrator_cli(
-        "execute",
-        "--task-dir",
-        str(task_dir),
-        "--route",
-        "small",
-        "--adapter",
-        "codex",
-        "--real",
-        env={"PATH": str(tmp_path / "empty-bin")},
-    )
-
-    assert result.returncode == 0
-    payload = json.loads(result.stdout)
-    assert payload["status"] == "failure"
-    assert payload["error"] == "codex binary not found on PATH; install/auth Codex CLI or omit --real to use the safe stub"
-
-
-def test_real_adapter_boundary_doc_defines_supported_slice_and_failure_modes() -> None:
-    doc = (ROOT / "docs" / "real-adapter-boundary.md").read_text(encoding="utf-8")
-
-    assert "Supported real execution slice" in doc
-    assert "Claude Code and Codex CLI" in doc
-    assert "Stub execution remains the default" in doc
-    assert "missing CLI" in doc
-    assert "auth failure" in doc
-    assert "non-zero exit" in doc
-    assert "malformed output" in doc
-
-
 def test_cli_execute_runs_current_stage_and_writes_output(tmp_path: Path) -> None:
     task_dir = _make_task_dir(tmp_path)
     execute_result = _run_orchestrator_cli("execute", "--task-dir", str(task_dir), "--route", "small", "--adapter", "codex")
@@ -594,161 +415,3 @@ def test_cli_medium_route_execute_flow_is_automated_end_to_end(tmp_path: Path) -
     resume_payload = json.loads(resume_result.stdout)
     assert resume_payload["current_stage"] == "execute"
     assert resume_payload["current_task_id"] == "task-1"
-
-
-def test_readme_examples_describe_manual_execution_flow() -> None:
-    readme = (ROOT / "README.md").read_text(encoding="utf-8")
-
-    assert "## Quickstart" in readme
-    assert "make validate" in readme
-    assert "make orchestrator-help" in readme
-    assert "make runtime-sample" in readme
-    assert "scripts/run_orchestrator.py init" in readme
-    assert "Other manual `run_orchestrator.py` commands mutate their target `--task-dir`" in readme
-    assert "scripts/run_orchestrator.py execute --task-dir" in readme
-    assert "advance --execute" in readme
-    assert "## Using ForgeFlow in Codex" in readme
-    assert "cp adapters/generated/codex/CODEX.md ./CODEX.md" in readme
-    assert "codex exec" in readme
-    assert "## Using ForgeFlow in Claude Code" in readme
-    assert "cp adapters/generated/claude/CLAUDE.md ./CLAUDE.md" in readme
-    assert "claude -p" in readme
-    assert "## Claude Code prompt templates" in readme
-    assert "### Small task template" in readme
-    assert "### Medium task template" in readme
-    assert "### Large / high-risk task template" in readme
-    assert "State the route you are using" in readme
-    assert "Do not merge spec-review and quality-review" in readme
-    assert "run`은 artifact/gate 기준으로 route 상태를 진행" in readme
-    assert "execute`는 현재 stage를 어댑터로 실행" in readme
-
-
-def test_runtime_sample_cli_uses_disposable_fixture_copy() -> None:
-    fixture_dir = ROOT / "examples" / "runtime-fixtures" / "small-doc-task"
-    tracked_files = {
-        path.relative_to(fixture_dir): path.read_text(encoding="utf-8")
-        for path in fixture_dir.rglob("*")
-        if path.is_file()
-    }
-
-    result = subprocess.run(
-        [
-            sys.executable,
-            "scripts/run_runtime_sample.py",
-            "--fixture-dir",
-            str(fixture_dir),
-            "--route",
-            "small",
-        ],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    assert result.returncode == 0
-    payload = json.loads(result.stdout)
-    assert payload["status"] == "completed"
-    assert payload["current_stage"] == "finalize"
-    assert payload["sample_source_fixture"] == "examples/runtime-fixtures/small-doc-task"
-    assert "sample_workspace" not in payload
-
-    for rel, original in tracked_files.items():
-        assert (fixture_dir / rel).read_text(encoding="utf-8") == original
-
-
-def test_runtime_sample_cli_rejects_non_directory_fixture(tmp_path: Path) -> None:
-    file_path = tmp_path / "not-a-dir.json"
-    file_path.write_text("{}", encoding="utf-8")
-
-    result = subprocess.run(
-        [
-            sys.executable,
-            "scripts/run_runtime_sample.py",
-            "--fixture-dir",
-            str(file_path),
-            "--route",
-            "small",
-        ],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    assert result.returncode == 1
-    assert result.stdout == ""
-    assert result.stderr.startswith("ERROR: fixture directory is not a directory:")
-
-
-def test_cli_reports_runtime_violations_without_tracebacks(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
-
-    for command_args, expected_error in [
-        (("run", "--task-dir", str(task_dir), "--route", "unknown"), "ERROR: unknown route: unknown"),
-        (
-            ("advance", "--task-dir", str(task_dir), "--route", "unknown", "--current-stage", "clarify"),
-            "ERROR: unknown route: unknown",
-        ),
-        (
-            ("step-back", "--task-dir", str(task_dir), "--route", "unknown", "--current-stage", "quality-review"),
-            "ERROR: unknown route: unknown",
-        ),
-        (("escalate", "--task-dir", str(task_dir), "--from-route", "unknown"), "ERROR: unknown route for escalation: unknown"),
-    ]:
-        result = _run_orchestrator_cli(*command_args)
-
-        assert result.returncode == 1
-        assert result.stdout == ""
-        assert result.stderr.startswith(expected_error)
-
-
-def test_medium_and_large_runtime_fixtures_run_end_to_end(tmp_path: Path) -> None:
-    fixtures_root = ROOT / "examples" / "runtime-fixtures"
-
-    for fixture_name, route_name, expected_stage in [
-        ("medium-refactor-task", "medium", "finalize"),
-        ("medium-plan-with-weak-verification", "medium", "finalize"),
-        ("large-migration-task", "large_high_risk", "long-run"),
-        ("large-approved-but-unsafe", "large_high_risk", "long-run"),
-    ]:
-        source_dir = fixtures_root / fixture_name
-        task_dir = tmp_path / fixture_name
-        subprocess.run(["cp", "-R", str(source_dir), str(task_dir)], check=True)
-
-        result = run_route(task_dir=task_dir, policy=load_runtime_policy(ROOT), route_name=route_name)
-
-        assert result["status"] == "completed"
-        assert result["current_stage"] == expected_stage
-
-
-def test_adherence_eval_cli_runs_valid_and_negative_fixtures() -> None:
-    result = subprocess.run(
-        [sys.executable, "scripts/run_adherence_evals.py"],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    assert result.returncode == 0
-    assert "ADHERENCE EVALS: PASS" in result.stdout
-    assert "small-doc-task" in result.stdout
-    assert "resume-small-task" in result.stdout
-    assert "medium-refactor-task" in result.stdout
-    assert "medium-plan-with-weak-verification" in result.stdout
-    assert "large-migration-task" in result.stdout
-    assert "large-approved-but-unsafe" in result.stdout
-    assert "missing-quality-approval" in result.stdout
-    assert "invalid-review-report" in result.stdout
-    assert "missing-run-state-before-spec-review" in result.stdout
-    assert "missing-eval-record-before-long-run" in result.stdout
-    assert "mixed-task-review-report" in result.stdout
-    assert "mixed-task-decision-log" in result.stdout
-    assert "checkpoint-gate-drift" in result.stdout
-    assert "future-gate-checkpoint-drift" in result.stdout
-    assert "completed-checkpoint-drift" in result.stdout
-    assert "medium-ledger-gate-drift" in result.stdout
-    assert "large-spec-quality-mismatch" in result.stdout
-    assert "large-session-state-stale-review-ref" in result.stdout
-    assert "medium-session-state-route-drift" in result.stdout
